requirements.txt
tweepy>=4.14.0
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.29.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-chroma>=0.1.0
chromadb>=0.4.0
